requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pandas==2.2.2
playwright==1.44.0
python-dotenv==1.0.1
schedule==1.2.1
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
